Serve public Esplora at /api and mempool-backend at /esplora
/api is the wallet/explorer electrs tree. Mempool-backend helpers
(/internal, /block-template) move to /esplora, which remains a
superset of the public routes so ESPLORA_REST_API_URL can point at it.
