+++ crates/rpc/README.md
@@ -17,8 +17,8 @@ no backend cargo feature (`g17_dependency_direction` proves both from
 
 ### 1. Protocol demuxing and authentication
 - **Transport Demuxing**: Private free function `serve_connection` (`crates/rpc/src/server.rs`) demuxes incoming HTTP requests before authentication:
-  - `GET`: If path starts with `/rest/*`, routed to unauthenticated `rest::route`; `/api/*` is unauthenticated Esplora (`esplora::route`). Unprefixed GET paths 404 so JSON-RPC keeps the listener root.
-  - `POST` (Esplora): Paths under `/api` are a closed Esplora namespace (`esplora::route_post`): they never fall through to JSON-RPC. That tree includes wallet `POST /api/tx` and mempool-backend `/api/internal/*`. Unprefixed POST paths (including `/tx`) fall through to JSON-RPC.
+  - `GET`: If path starts with `/rest/*`, routed to unauthenticated `rest::route`; `/api/*` is public Esplora and `/esplora/*` is the mempool-backend electrs superset (`esplora::route`). Unprefixed GET paths 404 so JSON-RPC keeps the listener root.
+  - `POST` (Esplora): Paths under `/api` and `/esplora` are closed namespaces (`esplora::route_post`): they never fall through to JSON-RPC. Wallet broadcast is `POST /api/tx`. Mempool-backend bulk paths live under `/esplora/internal/*`. Unprefixed POST paths (including `/tx`) fall through to JSON-RPC.
   - `POST` (JSON-RPC): Unhandled POST paths (such as `/`) fall through to JSON-RPC authentication. `serve_connection` in `crates/rpc/src/server.rs` calls `Auth::validate_header`, which is owned by `crates/rpc/src/auth.rs` (HTTP Basic / Cookie).
 - **JSON-RPC Framing & Protocol Versioning**: `JsonRpcVersion` (`crates/rpc/src/server.rs`) governs wire framing:
   - Requests with `"jsonrpc": "2.0"` use JSON-RPC 2.0 (`JsonRpcVersion::V2`): success responses emit `{"jsonrpc":"2.0","result":...,"id":...}` (HTTP 200), error responses emit `{"jsonrpc":"2.0","error":...,"id":...}` (HTTP 200), and requests omitting `id` are treated as notifications returning HTTP 204 No Content.
