+++ crates/rpc/src/server.rs
@@ -155,8 +155,9 @@ fn serve_connection(
 
         if request.method == "GET" {
             let (path, query) = split_path_query(&request.path);
-            // Routing policy is defined in docs/contracts/wallet-facing.md, WF-02;
-            // see that contract for the complete routing policy.
+            // `/rest/` is Core REST. `/api` and `/esplora` are Esplora
+            // directories (public electrs vs mempool-backend superset).
+            // Everything else 404s so JSON-RPC keeps `/`.
             let response = if path.starts_with("/rest/") {
                 crate::rest::route(handler.context(), path, query, rest_enabled)
             } else {
