+++ tools/btc-mempool-explorer/docker-compose.yaml
@@ -48,7 +48,7 @@ services:
     environment:
       MEMPOOL_NETWORK: "${BITCOIN_RS_NETWORK:-mainnet}"
       MEMPOOL_BACKEND: esplora
-      ESPLORA_REST_API_URL: http://node:8332/api
+      ESPLORA_REST_API_URL: http://node:8332/esplora
       CORE_RPC_HOST: node
       CORE_RPC_PORT: "8332"
       CORE_RPC_USERNAME: "${BITCOIN_RS_RPC_USER:-mempool}"
